+++ CHANGES.rst
@@ -18,6 +18,7 @@ Bugs fixed
 
 * #11668: Raise a useful error when ``theme.conf`` is missing.
   Patch by Vinay Sajip.
+* #11622: Fix ordering inside searchindex.js not being deterministic.
 
 Testing
 -------